drivers: Split uart_dev_dynamic_repo & uart_dev_static_repo
