+++ src/drivers/serial/uart_dev/Mybuild
@@ -1,8 +1,27 @@
 package embox.driver.serial
 
+@DefaultImpl(uart_repo_with_dynamic)
+abstract module uart_repo {
+}
+
+module uart_repo_with_dynamic extends uart_repo {
+	option number log_level = 1
+	option number uart_max_n = 4
+
+	source "uart_dev_dynamic_repo.c"
+
+	@NoRuntime depends embox.util.indexator
+	@NoRuntime depends embox.mem.pool
+}
+
+module uart_repo_static extends uart_repo {
+	option number log_level = 1
+
+	source "uart_dev_static_repo.c"
+}
+
 @DefaultImpl(core_tty)
 abstract module core {
-	option number uart_max_n = 4
 	option number log_level = 1
 
 	source "uart_dev.c"
@@ -10,10 +29,10 @@ abstract module core {
 	@IncludeExport(path="drivers/serial")
 	source "uart_dev.h"
 
-	@NoRuntime depends embox.util.indexator
 	@NoRuntime depends embox.util.dlist
 	@NoRuntime depends embox.util.log
-	@NoRuntime depends embox.mem.pool
+
+	depends uart_repo
 }
 
 module core_tty extends core {
